Sent a question to the authour
I have been working in the console of chrome dev tools. So that is why there are not to many changes to this. I can not get the page "all pies" to work correctly. I will come back to this turtorial when the author or someone gives me answers on this